Generative AI/LLM Projects/Gemini Model/YoutubeTranscriber/requirements.txt
streamlit
python-dotenv
google-generativeai
youtube-transcript-api
pathlib